Refactor MonthOverviewSummary component and related utilities
- Simplified MonthOverviewSummary by extracting utility functions and components.
- Created OverviewPeriodCard for rendering period-specific metrics.
- Introduced TooltipValue component for consistent tooltip rendering.
- Moved metric-related utility functions to monthOverviewSummaryUtils.js.
- Updated detailed comparison sections to use new utility functions for better readability and maintainability.
